profile as popup on chatcontainer
